[feat #106] Delivery 엔티티 생성 방법 변경
